Gnocchi-core finished, parquet input read in CLI still broken

+++ gnocchi-core/src/main/scala/net/fnothaft/gnocchi/models/GenotypeState.scala
@@ -24,7 +24,7 @@ case class GenotypeState(contig: String,
                          ref: String,
                          alt: String,
                          sampleId: String,
-                         genotypeState: Double) {
+                         genotypeState: Int) {
 
   def referenceAllele: (ReferenceRegion, String) = {
     (ReferenceRegion(contig, start, end), alt)
